+++ decorator.py
@@ -1,16 +1,20 @@
+import json
+import logging
 from functools import wraps
-from telegram import ChatAction
 
-LIST_OF_ADMINS = [12345678, 87654321]
+logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+                    level=logging.INFO)
+logger = logging.getLogger(__name__)
 
 
 def restricted(func):
     @wraps(func)
     def wrapped(bot, update, *args, **kwargs):
         user_id = update.effective_user.id
-        if user_id not in LIST_OF_ADMINS:
-            print("Unauthorized access denied for {}.".format(user_id))
-            return
+        with open('admin_list.txt', 'r') as admin:
+            if user_id not in admin:
+                print("Unauthorized access denied for {}.".format(user_id))
+                return
         return func(bot, update, *args, **kwargs)
 
     return wrapped
@@ -20,7 +24,7 @@ def log(func):
     @wraps(func)
     def wrapped(bot, update, *args, **kwargs):
         user = update.message.from_user
-        print("{}({}) called {}".format(user.first_name, user.id, func.__name__))
+        logger.info("{}({}) called {}".format(user.first_name, user.id, func.__name__))
         return func(bot, update, *args, **kwargs)
 
     return wrapped
@@ -39,3 +43,32 @@ def send_action(action):
         return command_func
 
     return decorator
+
+
+def choose_log(func):
+    @wraps(func)
+    def wrapped(bot, update, *args, **kwargs):
+        query = update.callback_query
+        user = query.from_user
+        logger.info("{}({}) choose {}".format(user.first_name, user.id, query.data))
+
+        return func(bot, update, *args, **kwargs)
+
+    return wrapped
+
+
+def test(func):
+    @wraps(func)
+    def wrapped(bot, update, *args, **kwargs):
+        user = update.message.from_user
+        with open('club.json', 'r+') as data:
+            members = json.load(data)
+            if user.id not in (m['UID'] for m in members):
+                members.append({"UID": user.id, "FirstName": user.first_name})
+                data.seek(0)
+                data.write(json.dumps(members))
+                data.truncate()
+
+        return func(bot, update, *args, **kwargs)
+
+    return wrapped
